split class
