+++ Card.py
@@ -1,0 +1,24 @@
+"""
+基本となるカードを定義するクラス
+rank:カードのランク, suit:カードのスート, value:カードから得る数
+"""
+
+
+class Card:
+    RANKS = ('A', '2', '3', '4', '5', '6', '7', '8', '9', '10', 'J', 'Q', 'K')
+    SUITS = ('Spade', 'Heart', 'Diamond', 'Club')
+
+    # 初期化
+    def __init__(self, rank, suit):
+        self.rank = rank
+        self.suit = suit
+        self.value = int(self.getvalue())
+
+    # ランクを数字に変換する
+    def getvalue(self):
+        if self.rank == 'A':
+            return 11
+        elif self.rank == 'J' or self.rank == 'Q' or self.rank == 'K':
+            return 10
+        else:
+            return self.rank
