+++ GamePlayer.py
@@ -1,0 +1,52 @@
+'''
+ゲーム参加者を表すスーパークラス
+プレイヤーとディーラーとで共通のアクションを定義していく
+'''
+
+
+class GamePlayer:
+
+    # 初期化関数
+    def __init__(self):
+        self.cards = []
+        self.total = 0
+        self.acetotal = 0
+        self.usedace = 0
+        self.burst = False
+        self.naturalbj = False
+        self.normalbj = False
+
+    # 　子オブジェクトから呼び出せる初期化関数
+    def initialize(self):
+        self.cards = []
+        self.total = 0
+        self.acetotal = 0
+        self.usedace = 0
+        self.burst = False
+        self.naturalbj = False
+        self.normalbj = False
+
+    # ゲームプレイヤーの手札の合計値を返す関数
+    def totalvalue(self):
+        i = 0
+        self.total = 0
+        self.acetotal = 0
+        cardnum = len(self.cards)
+
+        while i < cardnum:
+            if (self.cards[i].rank == 'A'):
+                self.acetotal += 1
+            self.total += self.cards[i].value
+            i += 1
+        self.total -= 10 * self.usedace
+
+        # プレイヤーのバースト判定の処理
+        if (self.total > 21):
+            if (self.acetotal - self.usedace > 0):
+                self.total -= 10
+                self.usedace += 1
+                if (self.total > 21):
+                    self.burst = True
+            else:
+                self.burst = True
+
